TypedFASTX

Documentation for TypedFASTX.

TypedFASTX.AAReader

TypedFASTX.AARecord

TypedFASTX.AAWriter

TypedFASTX.AbstractReader

TypedFASTX.AbstractReader

TypedFASTX.AbstractRecord

TypedFASTX.AbstractWriter

TypedFASTX.AbstractWriter

TypedFASTX.DNAReader

TypedFASTX.DNARecord

TypedFASTX.DNAWriter

TypedFASTX.QualityScores

TypedFASTX.RNAReader

TypedFASTX.RNARecord

TypedFASTX.RNAWriter

TypedFASTX.StringReader

TypedFASTX.StringRecord

TypedFASTX.StringWriter

FASTX.description

FASTX.identifier

FASTX.sequence

FASTX.sequence

TypedFASTX.AAReader — Type

AAReader

Alias for AbstractReader{LongAA}. Can be used for constructing TypedFASTAReader{LongAA} and TypedFASTQReader{LongAA} instances.

TypedFASTX.AARecord — Type

AARecord

Alias for AbstractRecord{LongAA}. Can be used for constructing TypedFASTA.Record{LongAA} and TypedFASTQ.Record{LongAA} instances.

TypedFASTX.AAWriter — Type

AAWriter

Alias for AbstractWriter{LongAA}. Can be used for constructing TypedFASTAWriter{LongAA} and TypedFASTQWriter{LongAA} instances.

TypedFASTX.AbstractReader — Type

AbstractReader{T}

Abstract type for typed FASTX readers. T is the type of the sequence in the records.

TypedFASTX.AbstractReader — Method

Constructor for AbstractReader{T} that looks at the file extension of the given path.

TypedFASTX.AbstractRecord — Type

AbstractRecord{T}

Abstract type for typed FASTX records.

TypedFASTX.AbstractWriter — Type

AbstractWriter{T}

Abstract type for typed FASTX writers. T is the type of the sequence in the records.

TypedFASTX.AbstractWriter — Method

Constructor for AbstractWriter{T} that looks at the file extension of the given path.

TypedFASTX.DNAReader — Type

DNAReader

Alias for AbstractReader{LongDNA{4}}. Can be used for constructing TypedFASTAReader{LongDNA{4}} and TypedFASTQReader{LongDNA{4}} instances.

TypedFASTX.DNARecord — Type

DNARecord

Alias for AbstractRecord{LongDNA{4}}. Can be used for constructing TypedFASTA.Record{LongDNA{4}} and TypedFASTQ.Record{LongDNA{4}} instances.

TypedFASTX.DNAWriter — Type

DNAWriter

Alias for AbstractWriter{LongDNA{4}}. Can be used for constructing TypedFASTAWriter{LongDNA{4}} and TypedFASTQWriter{LongDNA{4}} instances.

TypedFASTX.QualityScores — Type

QualityScores

A type for storing FASTQ quality with different encodings.

TypedFASTX.RNAReader — Type

RNAReader

Alias for AbstractReader{LongRNA{4}}. Can be used for constructing TypedFASTAReader{LongRNA{4}} and TypedFASTQReader{LongRNA{4}} instances.

TypedFASTX.RNARecord — Type

RNARecord

Alias for AbstractRecord{LongRNA{4}}. Can be used for constructing TypedFASTA.Record{LongRNA{4}} and TypedFASTQ.Record{LongRNA{4}} instances.

TypedFASTX.RNAWriter — Type

RNAWriter

Alias for AbstractWriter{LongRNA{4}}. Can be used for constructing TypedFASTAWriter{LongRNA{4}} and TypedFASTQWriter{LongRNA{4}} instances.

TypedFASTX.StringReader — Type

StringReader

Alias for AbstractReader{String}. Can be used for constructing TypedFASTAReader{String} and TypedFASTQReader{String} instances.

TypedFASTX.StringRecord — Type

StringRecord

Alias for AbstractRecord{String}. Can be used for constructing TypedFASTA.Record{String} and TypedFASTQ.Record{String} instances.

TypedFASTX.StringWriter — Type

StringWriter

Alias for AbstractWriter{String}. Can be used for constructing TypedFASTAWriter{String} and TypedFASTQWriter{String} instances.

FASTX.description — Method

description(record::AbstractRecord)::String

Get the entire description of record. Returns a String.

FASTX.identifier — Method

identifier(record::AbstractRecord)::SubString{String}

Get the identifier of record. The identifier is the first "word" of the description. Returns a SubString{String}.

FASTX.sequence — Method

sequence(record::AbstractRecord{T})::T

Returns the sequence of record.

FASTX.sequence — Method

sequence(T, record::AbstractRecord)::T

Returns the sequence of record, converted to type T.