TypedFASTX

Documentation for TypedFASTX.

TypedFASTX.NO_QUALITY

TypedFASTX.NoQuality

TypedFASTX.TypedRecord

TypedFASTX.error_prob_generator

TypedFASTX.error_probs

TypedFASTX.NO_QUALITY — Constant

A constant instance of the type NoQuality.

TypedFASTX.NoQuality — Type

Used in place of QualityScores to denote lack of quality scores.

TypedFASTX.TypedRecord — Type

TypedRecord{T}

TypedRecord{T}(id::S, seq::T, qual::Q = nothing) where {S, T, Q}

constructs a TypedRecord with given id, seq, and optional qual.

TypedRecord(id::S, seq::T, qual::Q = nothing) where {S, T, Q}`

constructs a TypedRecord with the type of seq inferred from the input.

TypedRecord{T}(id::S, seq::t, qual::Q = nothing) where {S, T, t, Q}`

constructs a TypedRecord with seq coerced to type T.

TypedRecord{T}(record::TypedRecord{t}) where {T, t}

converts an existing TypedRecord to a new TypedRecord with sequence type T.

A type to represent a generic biological sequence record.

Fields

identifier::String: The unique identifier for the sequence.

sequence::T: The biological sequence. It can be of type, including: String, LongDNA{4}, LongRNA{4} or LongAA.

quality::Union{Nothing, QualityScores}: The quality scores associated with the sequence. Nothing indicates absence of quality scores.

TypedFASTX.error_prob_generator — Method

error_prob_generator(record::TypedRecord)

Creates a generator of error probabilities for each character in the sequence.

TypedFASTX.error_probs — Method

error_probs(record::TypedRecord)

Creates a vector of error probabilities for each character in the sequence.