TypedFASTX

Documentation for TypedFASTX.

TypedFASTX.TypedRecord

TypedFASTX.error_prob_generator

TypedFASTX.error_probs

TypedFASTX.TypedRecord — Type

TypedRecord{T}

A type to represent a generic biological sequence record.

Fields

identifier::String: The unique identifier for the sequence.

sequence::T: The sequence itself. It can be of any type, including: String, LongDNA{4}, LongRNA{4} or LongAA.

quality::AbstractQuality: The quality scores associated with the sequence. Nothing indicates absence of quality scores.

TypedFASTX.error_prob_generator — Method

error_prob_generator(record::TypedRecord)

Creates a generator of error probabilities for each character in the sequence.

TypedFASTX.error_probs — Method

error_probs(record::TypedRecord)

Creates a vector of error probabilities for each character in the sequence.